[REFACTOR] #58 - data 폴더 리패키징
