+++ app/src/main/java/co/kr/bemyplan/data/entity/DailyContents.kt
@@ -1,4 +1,4 @@
-package co.kr.bemyplan.data
+package co.kr.bemyplan.data.entity
 
 data class DailyContents(
     val day: Int,
