+++ app/src/main/java/co/kr/bemyplan/data/entity/list/ContentModel.kt
@@ -1,4 +1,4 @@
-package co.kr.bemyplan.data.list
+package co.kr.bemyplan.data.entity.list
 
 data class ContentModel(
     val photo: Int,
