+++ app/src/main/java/co/kr/bemyplan/data/entity/purchase/before/ContentModel.kt
@@ -1,4 +1,4 @@
-package co.kr.bemyplan.data.purchase.before
+package co.kr.bemyplan.data.entity.purchase.before
 
 data class ContentModel(
     val photo: Int,
