+++ app/src/main/java/co/kr/bemyplan/ui/main/home/HomeAdapter.kt
@@ -4,7 +4,7 @@ import android.util.Log
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
-import co.kr.bemyplan.data.home.TempHomeData
+import co.kr.bemyplan.data.entity.main.home.TempHomeData
 import co.kr.bemyplan.databinding.ItemHomePlanBinding
 import com.bumptech.glide.Glide
 
@@ -12,7 +12,7 @@ class HomeAdapter(val itemClick: (TempHomeData) -> Unit) : RecyclerView.Adapter<
     val planList = mutableListOf<TempHomeData>()
 
     inner class HomeViewHolder(private val binding: ItemHomePlanBinding):RecyclerView.ViewHolder(binding.root){
-        fun onBind(data:TempHomeData){
+        fun onBind(data: TempHomeData){
             Log.d("bindtest", data.text)
             binding.tvHomeCommon.text = data.text
             Glide.with(binding.ivHomeCommon.context).load(data.img).into(binding.ivHomeCommon)
