+++ app/src/main/java/co/kr/bemyplan/ui/purchase/after/DailyContentsFragment.kt
@@ -2,15 +2,13 @@ package co.kr.bemyplan.ui.purchase.after
 
 import android.annotation.SuppressLint
 import android.os.Bundle
-import android.util.Log
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import co.kr.bemyplan.data.Spot
+import co.kr.bemyplan.data.entity.Spot
 import co.kr.bemyplan.databinding.FragmentDailyContentsBinding
 import co.kr.bemyplan.ui.purchase.after.adapter.DailyContentsAdapter
-import co.kr.bemyplan.util.clipTo
 
 class DailyContentsFragment : Fragment() {
     private lateinit var contentsAdapter: DailyContentsAdapter
