+++ app/src/main/java/co/kr/bemyplan/ui/purchase/after/adapter/DailyContentsAdapter.kt
@@ -1,21 +1,16 @@
 package co.kr.bemyplan.ui.purchase.after.adapter
 
 import android.annotation.SuppressLint
-import android.text.Layout
 import android.view.LayoutInflater
-import android.view.View
 import android.view.ViewGroup
 import androidx.databinding.BindingAdapter
 import androidx.databinding.ObservableArrayList
 import androidx.databinding.ViewDataBinding
 import androidx.recyclerview.widget.RecyclerView
-import co.kr.bemyplan.data.Spot
+import co.kr.bemyplan.data.entity.Spot
 import co.kr.bemyplan.databinding.ItemDailyContentsBinding
 import co.kr.bemyplan.databinding.ItemDailyRouteBinding
 import co.kr.bemyplan.util.clipTo
-import com.bumptech.glide.Glide
-import com.bumptech.glide.load.resource.bitmap.RoundedCorners
-import com.bumptech.glide.request.RequestOptions
 import java.lang.IllegalStateException
 
 class DailyContentsAdapter(private val viewType: Int): RecyclerView.Adapter<DailyContentsAdapter.DailyContentsViewHolder>() {
