+++ app/src/main/java/co/kr/bemyplan/ui/purchase/after/adapter/DayAdapter.kt
@@ -4,8 +4,7 @@ import android.annotation.SuppressLint
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
-import co.kr.bemyplan.R
-import co.kr.bemyplan.data.DailyContents
+import co.kr.bemyplan.data.entity.DailyContents
 import co.kr.bemyplan.databinding.ItemDayButtonBinding
 
 class DayAdapter: RecyclerView.Adapter<DayAdapter.DayViewHolder>() {
