+++ app/src/main/java/co/kr/bemyplan/ui/purchase/before/adapter/ContentAdapter.kt
@@ -5,7 +5,7 @@ import android.view.ViewGroup
 import androidx.databinding.DataBindingUtil
 import androidx.recyclerview.widget.RecyclerView
 import co.kr.bemyplan.R
-import co.kr.bemyplan.data.purchase.before.ContentModel
+import co.kr.bemyplan.data.entity.purchase.before.ContentModel
 import co.kr.bemyplan.databinding.ItemBeforePurchaseContentBinding
 
 class ContentAdapter: RecyclerView.Adapter<ContentAdapter.ContentViewHolder>() {
