+++ app/src/main/res/layout/activity_after_purchase.xml
@@ -6,7 +6,7 @@
     <data>
         <variable
             name="post"
-            type="co.kr.bemyplan.data.Post" />
+            type="co.kr.bemyplan.data.entity.Post" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
