+++ app/src/main/res/layout/fragment_my_plan.xml
@@ -7,10 +7,10 @@
         <import type="android.view.View"/>
         <variable
             name="profile"
-            type="co.kr.bemyplan.data.myplan.Profile" />
+            type="co.kr.bemyplan.data.entity.main.myplan.Profile" />
         <variable
             name="purchaseTour"
-            type="co.kr.bemyplan.data.myplan.PurchaseTour" />
+            type="co.kr.bemyplan.data.entity.main.myplan.PurchaseTour" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
