+++ app/src/main/res/layout/item_before_purchase_content.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="model"
-            type="co.kr.bemyplan.data.purchase.before.ContentModel" />
+            type="co.kr.bemyplan.data.entity.purchase.before.ContentModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
