+++ app/src/main/res/layout/item_daily_contents.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="spot"
-            type="co.kr.bemyplan.data.Spot" />
+            type="co.kr.bemyplan.data.entity.Spot" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
